Keep the viewed chat log visible when the software keyboard opens at a mid-scroll position #695
