+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "joshuafolkken-com",
-	"version": "0.266.0",
+	"version": "0.267.0",
 	"type": "module",
 	"private": true,
 	"scripts": {
